wow, mmap on HPUX is so amazingly broken.
It seems that the tdbtorture binaries from when we were still
attempting to use mmap on hpux are stuck in some wait state, and they
have left the binary in "text file busy", and the tdb
undeletable. This hack moves them out of the way, so we can start
doing builds again.


git-svn-id: file:///home/svn/build-farm/trunk@398 1e5ffdc8-eadd-0310-9daa-9cb4117fe24b