fix some build farm site loading error. We had build.pl run for *very* long time when the build log contained a line like
/Users/build/build_farm/build_test: line 793: vmstat: command not found

The regexp used here is so complex and costly that the processing of the
rest of the log being done by this expression took so long. We want to
match at this place on the buildfarm test results only, so making the
regexp stricter by prepending ^ before test: is sufficient and works
around the problem for now.

We should really make this regexp monsters more readable, document what
they do and try to make them more bullet proof.


git-svn-id: file:///home/svn/build-farm/trunk@882 1e5ffdc8-eadd-0310-9daa-9cb4117fe24b