Cleanups and fixes for new import-and-analyse.pl script
This also removes the build_stage_run table (not needed, best to
record the stages in the main record).

The uploads from the clients now appear in datadir/uploads, so that
they cannot touch datadir/oldrevs.

Also fixed build_status_from_html() to call build_status_from_string
with correct arguments (if you don't call it with $self-> then the
arguments are shifted to the left by one).

Andrew Bartlett


git-svn-id: file:///home/svn/build-farm/trunk@894 1e5ffdc8-eadd-0310-9daa-9cb4117fe24b