Use transactions properly in the build farm import script
In particular, this ensures we always do a rollback, which was why
many things were not being processed in the past, as it would
correctly refused to nest transactions.

To do this properly, we also had to explicitly use transactions in the
main hostdb.pm code as well.

Andrew Bartlett


git-svn-id: file:///home/svn/build-farm/trunk@906 1e5ffdc8-eadd-0310-9daa-9cb4117fe24b